requirements.txt
pyyaml
pick